backend/requirements.txt
twilio
python-dotenv
flask
pydub
playsound
pyaudio
sounddevice
soundfile
numpy
openai
scipy